refactor(library/URI): moves out of "/customTypes" directory
