+++ src/features/TextToImage/types/Output/Image.ts
@@ -1,4 +1,4 @@
-import type ImageURI from '@/library/customTypes/UniformResourceIdentifier/Data/Image';
+import type ImageURI from '@/library/UniformResourceIdentifier/Data/Image';
 
 interface TextToImage_Output_Image {
   uri: ImageURI;
