+++ src/library/UniformResourceIdentifier/Data/EncodingIdentifier/Any.ts
