+++ src/library/UniformResourceIdentifier/Data/EncodingIdentifier/all.ts
