+++ src/library/UniformResourceIdentifier/Data/EncodingIdentifier/exports.ts
