+++ src/library/UniformResourceIdentifier/Data/EncodingIdentifier/index.ts
