+++ src/library/UniformResourceIdentifier/Data/Image/Format/Any.ts
