+++ src/library/UniformResourceIdentifier/Data/Image/Format/all.ts
