+++ src/library/UniformResourceIdentifier/Data/Image/Format/exports.ts
