+++ src/library/UniformResourceIdentifier/Data/Image/Format/index.ts
