+++ src/library/UniformResourceIdentifier/Data/Image/index.ts
