+++ src/library/UniformResourceIdentifier/Data/index.ts
