+++ src/library/UniformResourceIdentifier/index.ts
